detector_block_fix/requirements.txt
kafka-python
websocket
websocket-client
pymongo
mysql-connector
